requirements.txt
torch
torchvision
open_clip_torch
Pillow
tqdm